feat: add composable for text editor setup
